move fp16 wrapper into Test_Utils.hpp
